requirements/runtime.txt
kiwisolver==1.4.5
lazy_loader==0.3
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.4
mdurl==0.1.2
-e git+https://github.com/open-mmlab/mmdetection.git@cfd5d3a985b0249de009b67d04f37263e11cdf3d#egg=mmdet
mmengine==0.10.3
model-index==0.1.11
networkx==3.1
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.34
ordered-set==4.1.0
oss2==2.17.0
packaging==23.2
pandas==2.0.3
pillow==10.2.0
platformdirs==4.2.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.20.0
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.4
PyWavelets==1.4.1
pywin32==306
PyYAML==6.0.1
qudida==0.0.4
regex==2023.12.25
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.4.2
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
shapely==2.0.2
six==1.16.0
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminaltables==3.1.10
threadpoolctl==3.2.0
tifffile==2023.7.10
tomli==2.0.1
tqdm==4.65.2
typing_extensions==4.9.0
tzdata==2024.1
urllib3==1.26.18
Werkzeug==3.0.1
yapf==0.40.2
zipp==3.17.0